requirements.txt
streamlit
pinecone
langchain
langgraph
langchain-pinecone
langchain-community
langchain-google-genai
python-dotenv